extract newDummyXxx functions
